feat(crossasset)!: convert model from float64 to float32
Convert the entire crossasset package (~2300 lines) from float64 to
float32 for model weights, forward/backward pass, and training data.
This enables GPU acceleration (GPU engines work with float32) and
halves memory bandwidth, addressing issues #381 and #384 about
crossasset training being 10-100x slower than PyTorch.

BREAKING CHANGE: Public API signatures changed:
- Forward([][]float64) -> Forward([][]float32)
- Train([][][]float64, ...) -> Train([][][]float32, ...)
- TrainGPU([][][]float64, ...) -> TrainGPU([][][]float32, ...)
- Predict([][]float64) -> Predict([][]float32)
- AttentionWeights([][]float64) -> AttentionWeights([][]float32)

Changes:
- cpuEngine: NewCPUEngine[float64] -> NewCPUEngine[float32]
- All weight arrays (headW, headB, layer weights): float64 -> float32
- All layer node types: Linear[float64] -> Linear[float32], etc.
- All tensor operations: tensor.New[float64] -> tensor.New[float32]
- Serialize format bumped to v2 (4 bytes/weight instead of 8)
- AdamW optimizer: AdamW[float64] -> AdamW[float32]
- timeseries/crossasset_engine.go: converts float64 windows to float32
  at the adapter boundary
- Test tolerances relaxed for float32 precision (gradient parity,
  attention weight sums)

Config fields (LearningRate, DropoutRate) and summary metrics
(TrainResult.Losses, FinalAccuracy) remain float64.

+++ crossasset/gpu_train_test.go
@@ -20,16 +20,16 @@ func testConfig() Config {
 	}
 }
 
-func testData(cfg Config, n int) ([][][]float64, [][]int) {
-	data := make([][][]float64, n)
+func testData(cfg Config, n int) ([][][]float32, [][]int) {
+	data := make([][][]float32, n)
 	labels := make([][]int, n)
 	for i := range n {
-		data[i] = make([][]float64, cfg.NSources)
+		data[i] = make([][]float32, cfg.NSources)
 		labels[i] = make([]int, cfg.NSources)
 		for s := range cfg.NSources {
-			data[i][s] = make([]float64, cfg.FeaturesPerSource)
+			data[i][s] = make([]float32, cfg.FeaturesPerSource)
 			for f := range cfg.FeaturesPerSource {
-				data[i][s][f] = float64(i*cfg.NSources*cfg.FeaturesPerSource+s*cfg.FeaturesPerSource+f) * 0.01
+				data[i][s][f] = float32(i*cfg.NSources*cfg.FeaturesPerSource+s*cfg.FeaturesPerSource+f) * 0.01
 			}
 			labels[i][s] = (i + s) % 3
 		}
